backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
safetensors
accelerate
